Proposal for a `min_weight` param to exp functions
One issue with using the current exp functions is that there's no `min_periods` parameter, meaning that initial results can be overly influenced by a couple of values.

This PR adds a `min_weight` parameter to the exp functions, which is the minimum weight that a value requires before producing a value.

Assuming that `alpha` is constant, this is equivalent to taking a rolling sum of `1` or `0`s and then passing the original result where the corresponding value of the rolling sum is greater than `min_weight/alpha`.

So for a series with a slower moving average (i.e. a lower alpha), we require more points to be present before producing a value. Similarly, a higher `min_weight` value will require more points to be present. If all values exist, then the value of `1/alpha` will asymptotically approach 1 — so a `min_weight` of 1 will never produce values.