Implement `group_nanstd` as a copy-paste
I hit small issues when wrapping functions with the `move_exp_nanstd`, and given these are getting more use, it's safer to implement it this way, and slightly faster, since the sqrt is fused. Not elegant!

(but I think we'd need something like Julia to be able to fuse the operations, and I'm not even sure Julia could do this given it's more than a simple broadcast...)