Add tests for segfaulting
In pursuit of https://github.com/numbagg/numbagg/issues/345, this adds some tests for whether we get segfaults.

It wasn't trivial to get working reproductions, and they depend on some delicate conditions, such as when the `__pycache__` paths were generated. Probably we can make the tests a bit smaller, but they do appear to work.

Unfortunately adding `_reload_parfors` doesn't seem to help; possibly because it's in the wrong place. I'll follow up with a PR with my current best attempt